Estucturando la app
